PATCH: MD5 htpasswd utility
The htpasswd utility that comes with thttpd currently uses the system
default crypt() one-way function for passwords. This is ridiculously
insecure as brute force password crackers can crack these files very
quickly. The attached patch modifies the included htpasswd utility to
use the MD5 algorithm version of the system crypt() function, if it is
available.

Advantages:

* Much, much more difficult to crack (better security, even if password
  files are stolen)
* thttpd will continue to authenticate the old style password files

Disadvantage:
* If your system does not have an MD5 version of the crypt() function,
  the default, "insecure" crypt will be used.

That said, a number of modern Unix systems (most?) support the MD5
version of crypt. I know that Linux, FreeBSD, OpenBSD all support it,
and have for some time. MacOSX does *not* support it.

A better patch would test the system crypt() facility to make sure that
MD5 passwords are supported as part of configure. If not, it would
compile in FreeBSD's implementation. Or else provide a fallback command
line switch However, I don't have time to play with configure and
friends.

Signed-off-by: Evan Jones <ejones@uwaterloo.ca>