Fix X-Forwarded-For when using IPv6, thanks Steve!
Thanks to https://blog.steve.fi/IPv6_and_thttpd.html I realized the
current X-Forwarded-For implementation was flawed.  What happened
was that REMOTE_ADDR for a CGI remained at 127.0.0.1 despite the
code for X-Forwarded-For actually did its job, or rather "did" ...

This patch changes the code to use the client's REAL IP with a
slightly ugly static char array per connection, which can now be
overridden by X-Forwarded-For.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>