Sub-directory support for .htpasswd and .htaccess files
This patch is an adaptation of thttpd-ng r18 which added recursive
lookup of .htpasswd files.  The code has been adapted to also handle
.htaccess files.  Originally based on a patch from Callum Gibson.

The idea is to seek for .htpasswd, or a .htaccess, file recursively in
parent directories until one is found.  This make it possible to set a
password file for a whole site subtree.  The -g command line option can
still be used to tell Merecat to use the global .htpasswd and .htaccess
files instead.  Documentation and source code updated appropriately.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>