Followup to f3e41a9: Tweaks to keep-alive support
This is probably one of the toughest patches I've ever made.  Not only
is the homebrew event code in old thttpd hard to follow, but it has now
become obvious to me that Merecat needs to be rewritten with a pure
event loop.

Also, it turns out that different browsers have different timeouts for
when to fall back to non-pipelined requests.  I.e., if we don't reply
quick enough on an already open socket, the browser will try to connect
again on the main server socket.  I guess multi-threaded, or pre-forked,
servers expect that to happen.

Anyway, I've verified now that the keep-alive implementation behaves as
a true persistent connection, funneling multiple requests over the same
socket.  However, I have no illusions that this can be improved upon
greatly by someone with a little more brains than me :)

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>