Initial HTTPS support
This patch adds HTTPS support using OpenSSL.  Currently it's not
possible to build without OpenSSL, this is planned for later.

The HTTPS support is very bare-bones atm.  Only a public certificate
file and a private key is supported, and it's only possible to use
with one host, not virtual hosts (unless you have a wildcard cert).

To enable, simply add the following to your merecat.conf:

    port     = 443
    ssl      = true
    certfile = /path/to/cert.pem
    keyfile  = /path/to/key.pem

Remember to set very strict permissions on the private key file!

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>