Remove command line options when .conf file support is enabled
This patch removes command line support for the following options when
Merecat is configured with confuse (.conf) file parser:

-c CGI
-d DIR
-g
-p PORT
-r
-s
-u USER
-v

Removing these options simplifies a lot of logic and makes it a lot
easier to further develop the code to support multiple servers {}
context from one daemon.  The [WEBROOT] and [HOSTNAME] options may
soon also be removed.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>