Log warnings and errors to stderr when running in foreground
This patch changes the default behavior of Merecat when running in the
foreground (-n).  Error messages, like failed to bind to port foo, are
now printed to stderr for convenience when running in the foreground.

To disable this behavior a the '-s' option has been changed to force the
use of syslog even when running in the syslog.  This means that the old
'-s' option, which controls the symlink behavior of merecat when built
without .conf file support, has moved to '-S'.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>