Fix regression from c1e6d0d and clean up sockaddr_t member names
When the real IP address support for X-Forwarded-For was introduced the
`sa_addr[]` was blindly added to the *union*.  Meaning, the original IP
address was overwritten, in the process losing any port information as
well.

This patch refactors the sockaddr_t to a struct with a union *member*.
We also drop the sa_stor member, which was never used, and rename the
member slightly for best practises.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>