Fix ordering bug, call openlog() before using syslog()
The .conf parser uses syslog() to log any errors it encounters.  To see
them on stderr the user can run merecat in the foreground, but with this
ordering bug openlog(), with LOG_PERROR, was not called until after the
conf_init() call, which can exit().  This can be very confusing.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>