fixing "make dist; make check"
After creating a tarball from a fresh git checkout with "make distcheck",
unpacking that tarball and running "make check", the tests would fail with

 fatal: making test-suite.log: failed to create php.trs
 fatal: making test-suite.log: failed to create php.log
 make[4]: *** [Makefile:487: test-suite.log] Error 1

This is due to tests/php.sh not being included in the tarball. Fixing it
by adding php.sh to EXTRA_DIST in tests/Makefile.am.

Fixes: github issue #31